Time: 0 ms (100.00%), Space: 7.5 MB (20.77%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -3,10 +3,10 @@ public:
 long long int binary_search(int n ){
     int s=0;
     int e=n;
-    long long int mid=s+(e-s)/2;
-    long long int ans=-1;
+    long int mid=s+(e-s)/2;
+    long int ans=-1;
     while(s<=e){
-        long long int sq=mid*mid;
+        long int sq=mid*mid;
         if(sq==n){
             return mid;
         }
